refactor(): 1단계 dls 실습 구현
