--- src/main/kotlin/BlackjackApplication.kt
@@ -1,6 +1,0 @@
-import controller.BlackJackController
-
-fun main() {
-    val blackjackController = BlackJackController()
-    blackjackController.run()
-}
