+++ src/main/kotlin/blackjack/BlackjackApplication.kt
@@ -1,0 +1,6 @@
+import blackjack.controller.BlackJackController
+
+fun main() {
+    val blackjackController = BlackJackController()
+    blackjackController.run()
+}
