+++ src/main/kotlin/blackjack/domain/card/Symbol.kt
@@ -1,0 +1,8 @@
+package blackjack.domain.card
+
+enum class Symbol(val symbol: String) {
+    DIAMOND("다이아몬드"),
+    SPADE("스페이드"),
+    HEART("하트"),
+    CLOVER("클로버")
+}
