+++ src/main/kotlin/blackjack/domain/gamer/Player.kt
@@ -1,0 +1,4 @@
+package blackjack.domain.gamer
+
+class Player(name: String, hand: Hand = Hand()) : Gamer(name, hand) {
+}
