+++ src/main/kotlin/blackjack/domain/gamer/Players.kt
@@ -1,0 +1,9 @@
+package blackjack.domain.gamer
+
+import blackjack.domain.card.Deck
+
+class Players(private val players: List<Player>) : List<Player> by players {
+    fun initStage(deck: Deck) {
+        repeat(2) { players.forEach { it.draw(deck.pop()) } }
+    }
+}
