--- src/main/kotlin/domain/card/Symbol.kt
@@ -1,8 +1,0 @@
-package domain.card
-
-enum class Symbol(val symbol: String) {
-    DIAMOND("다이아몬드"),
-    SPADE("스페이드"),
-    HEART("하트"),
-    CLOVER("클로버")
-}
