--- src/main/kotlin/domain/gamer/Player.kt
@@ -1,4 +1,0 @@
-package domain.gamer
-
-class Player(name: String, hand: Hand = Hand()) : Gamer(name, hand) {
-}
