+++ src/main/kotlin/dsl/Languages.kt
@@ -1,0 +1,7 @@
+package dsl
+
+class Languages(private val languages: MutableList<Pair<String, Int>> = mutableListOf()) :
+    List<Pair<String, Int>> by languages {
+
+    infix fun String.level(other: Int) = languages.add(Pair(this, other))
+}
