+++ src/main/kotlin/dsl/introduce.kt
@@ -1,0 +1,5 @@
+package dsl
+
+fun introduce(function: Person.() -> Unit): Person {
+    return Person().apply(function)
+}
