+++ src/test/kotlin/blackjack/domain/card/DeckTest.kt
@@ -1,0 +1,14 @@
+package blackjack.domain.card
+
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+
+class DeckTest {
+    @Test
+    internal fun name() {
+        val deck = Deck()
+        Assertions.assertThat(deck).hasSize(52)
+        deck.pop()
+        Assertions.assertThat(deck).hasSize(51)
+    }
+}
